requirements.txt
numpy==1.21.6
scikit-build==0.16.3
torch==1.13.0
torch-fidelity==0.3.0
torchmetrics==0.11.1
torchsummary==1.5.1
torchvision==0.14.0
torchviz==0.0.2
absl-py==1.3.0
aiohttp==3.8.1
aiosignal==1.3.1
albumentations==1.3.0
anyio==3.6.2
asttokens==2.1.0
async-timeout==4.0.2
asyncer==0.0.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
distro==1.8.0
executing==1.2.0
fastapi==0.80.0
filelock==3.8.0
filetype==1.1.0
flatbuffers==22.10.26
fonttools==4.38.0
frozenlist==1.3.3
gdown==4.5.1
geomloss==0.2.5
google-auth==2.14.1
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.50.0
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.2.1
imageio==2.22.4
imageio-ffmpeg==0.4.3
importlib-metadata==5.0.0
ipython==8.6.0
jedi==0.18.1
joblib==1.2.0
kiwisolver==1.4.4
kornia==0.6.9
line-profiler==4.0.2
llvmlite==0.39.1
lpips==0.1.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mpmath==1.2.1
multidict==6.0.2
munch==2.5.0
networkx==2.8.8
numba==0.56.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnxruntime==1.12.1
onnxruntime-gpu==1.12.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.32
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
Pygments==2.13.0
PyMatting==1.1.8
pyparsing==3.0.9
PySocks==1.7.1
pyspng==0.1.1
python-dateutil==2.8.2
python-multipart==0.0.5
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
rembg
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.7.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.1
starlette==0.19.1
sympy==1.11.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tqdm==4.64.0
traitlets==5.5.0
typing-extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0
segmentation_models_pytorch==0.3.2